Draft - MSW_Cygwin convenience improvements.

+++ ubiquitous_bash.sh
@@ -32,7 +32,7 @@ _ub_cksum_special_derivativeScripts_contents() {
 #export ub_setScriptChecksum_disable='true'
 ( [[ -e "$0".nck ]] || [[ "${BASH_SOURCE[0]}" != "${0}" ]] || [[ "$1" == '--profile' ]] || [[ "$1" == '--script' ]] || [[ "$1" == '--call' ]] || [[ "$1" == '--return' ]] || [[ "$1" == '--devenv' ]] || [[ "$1" == '--shell' ]] || [[ "$1" == '--bypass' ]] || [[ "$1" == '--parent' ]] || [[ "$1" == '--embed' ]] || [[ "$0" == "/bin/bash" ]] || [[ "$0" == "-bash" ]] || [[ "$0" == "/usr/bin/bash" ]] || [[ "$0" == "bash" ]] ) && export ub_setScriptChecksum_disable='true'
 export ub_setScriptChecksum_header='1891409836'
-export ub_setScriptChecksum_contents='3617790695'
+export ub_setScriptChecksum_contents='1526967800'
 
 # CAUTION: Symlinks may cause problems. Disable this test for such cases if necessary.
 # WARNING: Performance may be crucial here.
@@ -11132,6 +11132,12 @@ _closeLoop() {
 }
 
 _testCreateFS() {
+	if _if_cygwin
+	then
+		echo 'warn: accepted: cygwin: missing: mount: mkfs'
+		return 0
+	fi
+	
 	_mustGetSudo
 	
 	_getDep mkfs
@@ -11139,6 +11145,12 @@ _testCreateFS() {
 }
 
 _testCreatePartition() {
+	if _if_cygwin
+	then
+		echo 'warn: accepted: cygwin: missing: mount: parted'
+		return 0
+	fi
+	
 	_mustGetSudo
 	
 	_getDep parted
